requirements.txt
psycopg2-binary>=2.9.3
pymongo>=4.1.1
pandas>=1.4.2
SQLAlchemy>=1.4.36
python-dotenv>=0.20.0
tqdm>=4.64.0
pyyaml>=6.0
pytest>=7.1.2
